Fix access a non-object when no data provided by XenForo
Until now, the extension expected, that XenForo returns data in any
case. However, in some circumstances, the Xenforo API seems to return
a non-OK status code (e.g. 403) when trying to retrieve information about
the current user, even if an oauth_token is provided. This case
is now handled with this commit, with an error message that indicates,
that there was a problem with the external site and, if provided, an
error message.

Bug: T165517
Change-Id: I378610dd11b2fc0c9e1ee029f96039e3d305c9a8